==> 3_london_CAL_cont.Rmd <==

# AI to CAL

This file will first convert the measured access index to CAL using set frequency as PTAL and then analyse the CAL spatial distribution in relation to PTAL. 

## load packages 

```{r}
library(sf)
library(data.table)
library(dplyr)
library(tmap)
library(grid)
library(stringr)
library(spdep)
library(ggplot2)
library(tidyr)
library(patchwork)
library(ggplot2)
library(forcats)
```

## read in AI measured in previous file 

```{r}
ttm <- fread(file.path("london_ai.csv"))

ttm$from_id <- ttm$from_id%>%
  as.character(.)

summary(ttm)
```

## merge ptal data with AI

```{r}
origin_sf <- fread(file.path("input/origin.csv"))%>%
  st_as_sf(.,coords = c("lat", "lon"), crs = 4326)%>%
  select("id","AI2015","PTAL2015","geometry")

origin_sf$id <- origin_sf$id%>%
  as.character(.)

AI_and_PTAL <-ttm%>%
  left_join(.,origin_sf,by= join_by("from_id"=="id"))%>%
  st_as_sf(crs = 4326)
```

## convert AI to CAL 

It will first calculate the frequency of each band of PTAL and apply the same frequency to CAL band. 

```{r}
valid <- AI_and_PTAL%>%
  filter(AI2015 > 0) 
```

```{r}
# organise the data
valid2 <- valid%>%
  rename(AIpeak =AI, AIoffpeak = AI_off, AIPTAL = AI2015)%>%
  pivot_longer(
    cols = starts_with("AI"),
    names_to = "name",
    names_prefix = "AI",
    values_to = "AI",
    values_drop_na = TRUE
  )

# Most basic violin chart
plot <- ggplot(valid2, aes(x=name, y=AI, fill=name, color=name)) + 
  geom_violin()+
  theme_ipsum() +
  theme(legend.position="none")+
  coord_flip() +
      xlab("")

plot

ggsave("violin.png")
```

From the violin plot of CAL's AI at peak and offpeak and PTAL's AI, a positively skewed distribution is seen for all scenarios. Mode is smaller than mean and median. Most of the grids are of low AI with some extreme values. 

Since the maximum AI of the three scenarios are very different, PTAL's AI maxi is 200 while CAL's AI maxi is 500 for off peak and 550 for peak time. The conversion of AI to CAL will be using the same frequency of AI as of PTAL, rather than the fixed AI range as PTAL. 

```{r}
count_all <- length(valid$PTAL2015)

PTAL_summary <- AI_and_PTAL%>%
  group_by(PTAL2015)%>%
  summarise(average_AI = mean(AI2015), 
            count = n(), 
            pct = count/count_all)

```

### CAL conversion for peaktime values 

```{r}
valid_CAL <- AI_and_PTAL%>%
  filter(AI > 0)
  
count_all_cal <- length(valid_CAL$AI)

cal_summary <- PTAL_summary%>%
  mutate(n_cal = pct*count_all_cal)

AI_and_PTAL<-AI_and_PTAL%>%
  arrange(AI)

b1a <- nth(AI_and_PTAL$AI, round(cal_summary$n_cal[2]), na_rm = TRUE)
b1b <- nth(AI_and_PTAL$AI, round(sum(cal_summary$n_cal[2:3])), na_rm = TRUE)
b2 <- nth(AI_and_PTAL$AI, round(sum(cal_summary$n_cal[2:4])), na_rm = TRUE)
b3 <- nth(AI_and_PTAL$AI, round(sum(cal_summary$n_cal[2:5])), na_rm = TRUE)
b4 <- nth(AI_and_PTAL$AI, round(sum(cal_summary$n_cal[2:6])), na_rm = TRUE)
b5 <- nth(AI_and_PTAL$AI, round(sum(cal_summary$n_cal[2:7])), na_rm = TRUE)
b6a <- nth(AI_and_PTAL$AI, round(sum(cal_summary$n_cal[2:8])), na_rm = TRUE)
b6b <- nth(AI_and_PTAL$AI, round(sum(cal_summary$n_cal[2:9])), na_rm = TRUE)

```

```{r}
CAL_and_PTAL_peak <- AI_and_PTAL%>%
  mutate(CAL = case_when(AI > 0 & AI <= b1a  ~ "1a",
                         AI > b1a & AI <= b1b  ~ "1b",
                         AI > b1b & AI <= b2   ~ "2",
                         AI > b2  & AI <= b3   ~ "3",
                         AI > b3  & AI <= b4   ~ "4",
                         AI > b4  & AI <= b5   ~ "5",
                         AI > b5  & AI <= b6a  ~ "6a",
                         AI > b6a   ~ "6b",
                         is.na(AI)~ "0",
                         TRUE ~ NA))
```

```{r}
count_all_cal <- length(CAL_and_PTAL_peak$CAL) - sum(is.na(CAL_and_PTAL_peak$CAL))

CAL_summary <- CAL_and_PTAL_peak%>%
  group_by(CAL)%>%
  summarise(average_AI = mean(AI), count = n(), 
            pct = count/count_all_cal)
```

### CAL conversion for off-peak values 

```{r}
valid_CAL_off <- AI_and_PTAL%>%
  filter(AI_off > 0)
  
count_all_cal_off <- length(valid_CAL_off$AI_off)

cal_summary_off <- PTAL_summary%>%
  mutate(n_cal = pct*count_all_cal_off)

AI_and_PTAL<-AI_and_PTAL%>%
  arrange(AI_off)

b1a <- nth(AI_and_PTAL$AI_off, round(cal_summary_off$n_cal[2]), na_rm = TRUE)
b1b <- nth(AI_and_PTAL$AI_off, round(sum(cal_summary_off$n_cal[2:3])), na_rm = TRUE)
b2 <- nth(AI_and_PTAL$AI_off, round(sum(cal_summary_off$n_cal[2:4])), na_rm = TRUE)
b3 <- nth(AI_and_PTAL$AI_off, round(sum(cal_summary_off$n_cal[2:5])), na_rm = TRUE)
b4 <- nth(AI_and_PTAL$AI_off, round(sum(cal_summary_off$n_cal[2:6])), na_rm = TRUE)
b5 <- nth(AI_and_PTAL$AI_off, round(sum(cal_summary_off$n_cal[2:7])), na_rm = TRUE)
b6a <- nth(AI_and_PTAL$AI_off, round(sum(cal_summary_off$n_cal[2:8])), na_rm = TRUE)
b6b <- nth(AI_and_PTAL$AI_off, round(sum(cal_summary_off$n_cal[2:9])), na_rm = TRUE)

```

```{r}
CAL_and_PTAL_off <- AI_and_PTAL%>%
  mutate(CAL_off = case_when(AI_off > 0 & AI_off <= b1a  ~ "1a",
                         AI_off > b1a & AI_off <= b1b  ~ "1b",
                         AI_off > b1b & AI_off <= b2   ~ "2",
                         AI_off > b2  & AI_off <= b3   ~ "3",
                         AI_off > b3  & AI_off <= b4   ~ "4",
                         AI_off > b4  & AI_off <= b5   ~ "5",
                         AI_off > b5  & AI_off <= b6a  ~ "6a",
                         AI_off > b6a   ~ "6b",
                         is.na(AI_off)~ "0",
                         TRUE ~ NA))
```

```{r}
count_all_cal_off <- length(CAL_and_PTAL_off$CAL_off) - sum(is.na(CAL_and_PTAL_off$CAL_off))

CAL_summary_off <- CAL_and_PTAL_off%>%
  group_by(CAL_off)%>%
  summarise(average_AI = mean(AI_off), count = n(), 
            pct = count/count_all_cal_off)
```

```{r}
PTAL_box <- 
ggplot(AI_and_PTAL, aes(x = factor(PTAL2015), fill = factor(PTAL2015))) +
    geom_bar(color="black")+
  scale_fill_manual(values=c('white','#134480',"#1b71be","#3cb9e8", "#8dcc5b", "#ffeb3d", "#ffc08f", "#ff2e17","#8b1b16","black"))+
      theme_ipsum() +
        xlab("PTAL") + 
    ylab("frequency") +
  theme(legend.position="none") 

CALpeak_box <- 
ggplot(CAL_and_PTAL_peak, aes(x = factor(CAL), fill = factor(CAL))) +
    geom_bar(color="black")+
  scale_fill_manual(values=c('white','#134480',"#1b71be","#3cb9e8", "#8dcc5b", "#ffeb3d", "#ffc08f", "#ff2e17","#8b1b16","black"))+  
  theme_ipsum() +
        xlab("CAL(peak)") + 
    ylab("frequency") +
  theme(legend.position="none") 

CALoffpeak_box <- 
ggplot(CAL_and_PTAL_off, aes(x = factor(CAL_off), fill = factor(CAL_off))) +
    geom_bar(color="black")+
  scale_fill_manual(values=c('white','#134480',"#1b71be","#3cb9e8", "#8dcc5b", "#ffeb3d", "#ffc08f", "#ff2e17","#8b1b16","black"))+ 
    theme_ipsum() +
       xlab("CAL(offpeak)") + 
    ylab("frequency") +
  theme(legend.position="none") 

```

This figure combines the violin plot and box plots above with patchwork. 

```{r}
# One plot at the left and two at the right
plot <- ggplot(valid2, aes(x=name, y=AI, fill=name, color=name)) + 
  geom_violin()+
  theme_ipsum() +
  theme(legend.position="none")+
  coord_flip() +
      xlab("")

plot + (PTAL_box / CALpeak_box / CALoffpeak_box)  +
    plot_layout(ncol = 2,
              widths = c(6, 4)) + 
  plot_annotation(title = "Violin plot of AI and and bar plots of Accessibility Levels",
             theme = theme(plot.title = element_text(hjust = 0.5), text = element_text(size = 20), axis.text = element_text(size = 16), legend.title = element_text(size = 20)))
```

```{r}
ggsave("combine.png", height = 8, width = 12)
```

## add grid polygon for mapping

```{r}
grid <- st_read(here::here("data/PTAL_grid_cell_2015/PTAL_100m_Grid_Cells.shp"))%>%
  st_transform(., 4326)%>%
  select("ID")

grid$ID <- grid$ID%>%
  as.character()

CAL_and_PTAL_off <-CAL_and_PTAL_off%>%
  st_drop_geometry()%>%
  select("from_id","CAL_off")

CAL_and_PTAL_peak <-CAL_and_PTAL_peak%>%
  st_drop_geometry()

CAL_and_PTAL_all <-CAL_and_PTAL_peak%>%
  left_join(.,CAL_and_PTAL_off,by= join_by("from_id"=="from_id"))
  
CAL_and_PTAL_all <-grid%>%
  left_join(.,CAL_and_PTAL_all,by= join_by("ID"=="from_id"))%>%
  st_as_sf()
```

## summarise the range of each band

```{r}
count_cal_off <- length(CAL_and_PTAL_all$CAL_off) - sum(is.na(CAL_and_PTAL_all$CAL_off))
count_cal_peak <- length(CAL_and_PTAL_all$CAL) - sum(is.na(CAL_and_PTAL_all$CAL))

CAL_band_off <- CAL_and_PTAL_all%>%
  st_drop_geometry()%>%
  group_by(CAL_off)%>%
  summarise(mini = min(AI_off), maxi = max(AI_off), frq = n()/count_cal_off)

CAL_band_peak <- CAL_and_PTAL_all%>%
  st_drop_geometry()%>%
  group_by(CAL)%>%
  summarise(mini = min(AI),maxi = max(AI),frq = n()/count_cal_peak)
```

```{r}
st_write(CAL_band_peak, "cal_band_peak.csv")
st_write(CAL_band_off, "cal_band_offpeak.csv")
```
Calculate the changes of CAL levels at peak and at off-peak. 

```{r}
CAL_and_PTAL_all <- CAL_and_PTAL_all%>%
  mutate(CAL_num = case_when(CAL == "1a" ~ 1,
                         CAL == "1b"~ 2,
                         CAL == "2" ~ 3,
                         CAL ==  "3" ~ 4,
                         CAL ==  "4" ~ 5,
                         CAL == "5" ~ 6,
                         CAL ==  "6a" ~ 7,
                         CAL == "6b" ~ 8,
                        TRUE~ 0))%>%
  mutate(CAL_num_off = case_when(CAL_off == "1a" ~ 1,
                         CAL_off == "1b"~ 2,
                         CAL_off == "2" ~ 3,
                         CAL_off ==  "3" ~ 4,
                         CAL_off ==  "4" ~ 5,
                         CAL_off == "5" ~ 6,
                         CAL_off ==  "6a" ~ 7,
                         CAL_off == "6b" ~ 8,
                        TRUE~ 0))%>%
  mutate(diff = case_when(CAL != 0 & CAL_off != 0 ~ CAL_num - CAL_num_off,
                          TRUE ~ NA))
```

```{r}
CAL_and_PTAL_diff <- CAL_and_PTAL_all%>%
  filter(diff != 0)

CAL_and_PTAL_diff$diff <- CAL_and_PTAL_diff$diff %>%
  as.numeric()
```

## map the output

```{r}
london <- st_read(here::here("data/London_Boroughs.gpkg"))%>%
  st_transform(., 4326)%>%
  mutate(inner_outer = case_when(ons_inner == "F" ~ "Outer London",
         TRUE ~ "Inner London"))
```

```{r}
tmap_mode("plot")

map_AI_cal =
  tm_shape(CAL_and_PTAL_all) +
  tm_polygons(col = "AI",palette="-viridis",
          n = 6,
          style = "quantile",
          scale = 3, 
          showNA = FALSE,
          colorNA = 'black',
          legend.hist = TRUE,
          title = "Access Index (CAL, peak)",
          border.alpha = 0) +
tm_shape(london) +
  tm_polygons(col = NA, 
              alpha = 0)+ 
tm_layout(title="AI",
             frame = FALSE,
             legend.outside = TRUE,
             legend.text.size = 1,
             legend.hist.size = 1,
             legend.hist.width = 0.75)

map_AI_cal

map_AI_cal_off =
  tm_shape(CAL_and_PTAL_all) +
  tm_polygons(col = "AI_off",palette="-viridis",
          n = 6,
          style = "quantile",
          scale = 3, 
          showNA = FALSE,
          colorNA = 'black',
          legend.hist = TRUE,
          title = "Access Index (CAL, off-peak)",
          border.alpha = 0) +
tm_shape(london) +
  tm_polygons(col = NA, 
              alpha = 0)+ 
tm_layout(title="AI",
             frame = FALSE,
             legend.outside = TRUE,
             legend.text.size = 1,
             legend.hist.size = 1,
             legend.hist.width = 0.75)

map_AI_cal_off

map_AI_ptal =
  tm_shape(CAL_and_PTAL_all) +
  tm_polygons(col = "AI2015",palette="-viridis",
          n = 6,
          style = "quantile",
          scale = 3, 
          showNA = FALSE,
          colorNA = 'black',
          legend.hist = TRUE,
          title = "Access Index (PTAL)",
          border.alpha = 0) +
tm_shape(london) +
  tm_polygons(col = NA, 
              alpha = 0)+ 
tm_layout(title="AI",
             frame = FALSE,
             legend.outside = TRUE,
             legend.text.size = 1,
             legend.hist.size = 1,
             legend.hist.width = 0.75)


map_AI_ptal
```

```{r}
tmap_save(map_AI_cal, 'london_ai_cal.png')
tmap_save(map_AI_cal_off, 'london_ai_off_cal.png')
tmap_save(map_AI_ptal, 'london_ai_ptal.png')
```

The above map shows the Access Index values, before converted into the CAL bands. Extremely high values tend to concentrate in inner London...Unlike PTAL's AI exhibiting clusters of low values (coldspots) in central locations, CAL's AI is generally smoothly transitioning to low values from inner to outer London locations with limited coldspots within the centre. 



```{r}
COLOR = c('white','#134480',"#1b71be","#3cb9e8", "#8dcc5b", "#ffeb3d", "#ffc08f", "#ff2e17","#8b1b16")
# same color palette as PTAL
Labels = c('0: missing values', '1a: least accessible','1b','2','3','4','5','6a','6b: most accessible')

tmap_mode("plot")

map_cal =
  tm_shape(CAL_and_PTAL_all) +
  tm_polygons(col = "CAL",
              palette=COLOR,
              showNA = FALSE,
              colorNA = 'white',
              title = "CAL",
              border.alpha = 0,
              labels= Labels
              ) +
tm_shape(london) +
  tm_polygons(col = NA, 
              alpha = 0)+ 
tm_layout(bg.color = NA, 
          legend.show = FALSE, 
          main.title="A) Cycle Access Level (peak time)",
          main.title.size = .8,
          frame = FALSE)

map_cal


map_cal_off =
  tm_shape(CAL_and_PTAL_all) +
  tm_polygons(col = "CAL_off",
              palette=COLOR,
              showNA = FALSE,
              colorNA = 'white',
              title = "CAL",
              border.alpha = 0,
              labels= Labels) +
tm_shape(london) +
  tm_polygons(col = NA, 
              alpha = 0)+ 
tm_layout(bg.color = NA, 
          legend.show = FALSE, 
          main.title="B) Cycle Access Level (off-peak time)",
          main.title.size = .8,
          frame = FALSE)

map_cal_off

map_ptal =
  tm_shape(CAL_and_PTAL_all) +
  tm_polygons(col = "PTAL2015",
              palette=COLOR,
              showNA = FALSE,
              colorNA = 'white',
              title = "PTAL",
              border.alpha = 0,
              labels= Labels) +
tm_shape(london) +
  tm_polygons(col = NA, 
              alpha = 0)+ 
tm_layout(bg.color = NA, 
          legend.show = FALSE, 
          main.title="D) Public Transport Access Level",
          main.title.size = .8,
          frame = FALSE)

map_ptal

legend =
   tm_shape(CAL_and_PTAL_all) +
    tm_polygons("PTAL2015",
                palette=COLOR,
                showNA = FALSE,
                labels= Labels,
                title = "CAL and PTAL") +
    tm_layout(legend.only = TRUE, legend.position=c(0.2,0),asp=0.1)

legend


map_diff =
  tm_shape(CAL_and_PTAL_diff) +
  tm_polygons(col = "diff",palette="RdBu", 
          colorNA = 'white',
          midpoint = 0,
          showNA = FALSE,
          legend.hist = TRUE,
          title = "Difference in CAL",
          border.alpha = 0) +
tm_shape(london) +
  tm_polygons(col = NA, 
              alpha = 0)+ 
tm_layout(bg.color = NA, 
          frame = FALSE,
          legend.show = FALSE, 
          main.title="C) Peak and off-peak difference in CAL",
          main.title.size = .8)

map_diff


legend_diff =
   tm_shape(CAL_and_PTAL_diff) +
    tm_polygons("diff",
                palette="RdBu",
                midpoint = 0,
                showNA = FALSE,
                title = "Difference in CAL") +
    tm_scale_bar(breaks = c(0, 2, 4),position=c(0.3,0.04), text.size=0.6)+
    tm_compass(north=0, position=c(0.2,0.1))+
    tm_layout(legend.only = TRUE, legend.position=c(0.2,0.31),asp=0.1) +
    tm_credits("Data source: OSM, GLA, TfL. R package: r5r.", position=c(0.2,0.0), fontface = "italic")

legend_diff
  
```

```{r}
t=tmap_arrange(map_cal, map_cal_off, map_diff, map_ptal, legend_diff, legend, ncol=2)
t

```

```{r}
tmap_save(map_cal, 'london_cal_peak.png')
tmap_save(map_cal_off, 'london_cal_offpeak.png')
tmap_save(map_ptal, 'london_ptal.png')

tmap_save(t, 'all.png')

```

The output shows the converted result using the same frequency as of PTAL bands, highlighting the most accessible locations.PTAL map shows clear clusters of high values throughout London while CAL's high value cluster is more like one massive cluster at inner London.   

## export the data into csv

```{r}
CAL_and_PTAL_all <- CAL_and_PTAL_all%>%
  select("ID","AI","AI_off","AI2015", "PTAL2015","CAL","CAL_off" , "diff","geometry")

# drop Z dimension to export the data
new <- st_zm(CAL_and_PTAL_all, drop=T, what='ZM')

st_write(new, "london_cal.gpkg", driver = "GPKG",layer_options = "OVERWRITE=true")
```

# Spatial autocorrelation 

Areas with high accessibility can be also influential to their surrounding locations. People living in the grid close to high AI values can also enjoy the good proximity and accessibility through sustainable transport. Hence, spatial autocorrelation of the AI values will be assessed to see if there is any cluster of such places that could be the ideal locations for cycle infrastructure investment and new housing development projects.  

## Moran's I

- Spatial weight matrix 

```{r}
new_dropna <-new %>%
  filter(AI > 0)
```

```{r}
coordsW <- new_dropna%>%
  st_centroid()%>%
  st_geometry()%>%
  st_transform(.,27700)

new_27700 <- new_dropna%>%
  st_transform(.,27700)

nb <- new_27700%>%
  poly2nb(., queen=TRUE) #produce a neighbours list based on regions with contiguous boundaries

#plot them
plot(nb, st_geometry(coordsW), col="red")
```

```{r}
listw <- nb2listw(nb, zero.policy = TRUE)
```

- Computing the Global Moran’s I

```{r}
globalMoran <- moran.test(new_27700$AI, listw, zero.policy=TRUE,na.action=na.omit)
globalMoran

I <- globalMoran[["estimate"]][["Moran I statistic"]]
I
pvalue <- globalMoran[["p.value"]]
pvalue
```

Global morans I statistic is 0.9368 with p value close to 0. The p-value is statistically significant, and the z-score is positive. This reveals that the Access index is not randomly distributed among the features in Camden and the AI is clustered with similar values. 

To see the location of the hot spots, local morans i can be measured for each grid. This project includes codes that are adapted from https://rpubs.com/quarcs-lab/spatial-autocorrelation. 

Local spatial autocorrelation

```{r}
moran <- moran.plot(new_27700$AI, listw = nb2listw(nb, style = "W", zero.policy = TRUE), zero.policy = TRUE, na.action= na.omit)
```

- Compute local Moran

```{r}
local <- localmoran(x = new_27700$AI, listw = nb2listw(nb, style = "W", zero.policy = TRUE), zero.policy = TRUE, na.action= na.omit)

moran.map <- cbind(new_27700, local)

tm_shape(moran.map) +
  tm_fill(col = "Ii",
          style = "quantile",
          title = "local moran statistic") +
  tm_layout(title="local morans I",
             frame = FALSE,
             legend.outside = TRUE,
            legend.text.size = 1,
             legend.hist.size = 1,
             legend.hist.width = 0.75)
```

- Plot LISA clusters

```{r}
quadrant <- vector(mode="numeric",length=nrow(local))

# centers the variable of interest around its mean
m.AI <- new_27700$AI - mean(new_27700$AI)     

# centers the local Moran's around the mean
m.local <- local[,1] - mean(local[,1])    

# significance threshold
signif <- 0.1 

# builds a data quadrant
quadrant[m.AI >0 & m.local>0] <- "4"  
quadrant[m.AI <0 & m.local<0] <- "1"      
quadrant[m.AI <0 & m.local>0] <- "2"
quadrant[m.AI >0 & m.local<0] <- "3"
quadrant[local[,5]>signif] <- "0"   


quadrant.map <- cbind(new_27700, quadrant)

quadrant.map <- quadrant.map %>%
  mutate(cluster = case_when(quadrant == "0" ~ "insignificant",
                             quadrant == "1" ~ "low-low",
                             quadrant == "2" ~ "low-high",
                             quadrant == "3" ~ "high-low",
                             quadrant == "4" ~ "high-high"))

cluster.map <- quadrant.map%>%
  select("geometry","cluster")
```

```{r}
brks <- c("0","1","2","3","4")
colors <- c("white","blue",rgb(0,0,1,alpha=0.4),rgb(1,0,0,alpha=0.4),"red")
Palette <- c("white", "blue", "lightblue", "pink", "red" )

Labels = c("insignificant", "low-low", "low-high", "high-low" ,"high-high")
  
tmap_mode("plot")

map_cluster = 
  tm_shape(london) +
  tm_polygons(col = NA, 
              alpha = 0)+ 
tm_shape(quadrant.map) +
  tm_polygons(col="quadrant",
          breaks = brks,
          palette=Palette,
          showNA = FALSE,
          colorNA = 'white',
          title = "cluster",
          border.alpha = 0,
          labels= Labels) +
tm_shape(london) +
  tm_polygons(col = NA, 
              alpha = 0)+ 
tm_layout(bg.color = NA, 
            legend.position = c(.85,0.05),
            main.title="LISA cluster map",
            main.title.size = 1,
             frame = FALSE,
             legend.text.size = 0.7)+
  tm_compass(position=c(0,0.1)) + 
  tm_scale_bar(breaks = c(0, 2, 4),position=c(0,0.03)) +
  tm_credits("Data source: OSM, GLA, TfL. R package: r5r.", fontface = "italic", position=c(0,0))

map_cluster
```

This output shows that inner London as a large cluster generally has high CAL surrounded by also high CAL. Outer London on the other hand has lower CAL and surrounded by low CAL grids. Good local connectivity is concentrated in central locations while outer boroughs obtain poorer local connectivity.Those areas are more likely to be car dependent as the local connectivity through walking, cycling and public transit is poor. The middle ring between inner and outer London clusters shows not significant result. A mixture of high and low CAL is present in this ring, implying high potential for planners to direct less infrastructure into the areas to gain higher return in improving local connectivity. 

```{r}
tmap_save(map_cluster, 'LISA_cluster.png')
```

# CAL per borough

Average CAL will then be measured per borough to identify case study area in this project. And compare between the inner and outer london boroughs. 

Add borough and lsoa info to each grid. 

```{r}
CAL_and_PTAL_all <- st_read(here::here("london_cal.gpkg"))


lsoa <- st_read(here::here("data/lsoa_geo_shp/LSOA_2011_London_gen_MHW.shp"))%>%
  st_transform(.,4326)%>%
  dplyr::select("LSOA11CD","LSOA11NM","LAD11CD","LAD11NM", "geometry")

lsoa22 <- read.csv(here::here("data/lsoa11to22.csv"),sep = ",",  header = TRUE, 
                         encoding = "latin1")%>%
  filter(LSOA11CD != LSOA21CD)%>%
  dplyr::select("LSOA11CD", "LSOA21CD")

lsoa_new <- lsoa%>%
  left_join(., lsoa22, by = join_by("LSOA11CD" == "LSOA11CD"))%>%
  mutate(code = case_when(!is.na(LSOA21CD) ~ LSOA21CD,
                          TRUE~ LSOA11CD))%>%
  select("code", "LSOA11NM", "LAD11NM", "geometry")

lsoa_new <- lsoa_new%>%
  unique()

sf_use_s2(FALSE)

CAL_and_PTAL_lsoa_grid <-  st_join(lsoa_new, CAL_and_PTAL_all)
```

```{r}
qtm(lsoa_new)
```

```{r}
CAL_and_PTAL_lsoa_grid <- CAL_and_PTAL_lsoa_grid%>%
  unique()
```

- calculate average AI of each lsoa 

```{r}
CAL_and_PTAL_lsoa <- CAL_and_PTAL_lsoa_grid %>%
  mutate(AI_mean = (AI+AI_off)/2)%>%
  group_by(code)%>%
  summarize(average_AI=mean(AI_mean,na.rm=TRUE))
```
```{r}
brent <- london%>%
  filter(name == "Brent")
brom <- london%>%
  filter(name == "Bromley")
```

```{r}
tmap_mode("plot")

map_mean_ai = tm_shape(CAL_and_PTAL_lsoa) +
  tm_fill(col="average_AI",
          palette="-viridis",
          title = "average AI",
          border.alpha = 0) +
tm_shape(london) +
  tm_polygons(col = NA, 
              alpha = 0)+ 
tm_shape(brent) +
  tm_polygons(col = NA, 
              border.col = "red",
              alpha = 0)+ 
  tm_text("name")+
tm_shape(brom) +
  tm_polygons(col = NA, 
              border.col = "red",
              alpha = 0)+ 
  tm_text("name")+
  tm_layout(bg.color = NA, 
            legend.position = c(.85,0.05),
            main.title="Average Access Index map",
            main.title.size = 1,
             frame = FALSE,
             legend.text.size = 0.7)+
  tm_compass(position=c(0,0.1)) + 
  tm_scale_bar(breaks = c(0, 2, 4),position=c(0,0.03)) +
  tm_credits("Data source: OSM, GLA, TfL. R package: r5r.", fontface = "italic", position=c(0,0))

map_mean_ai
```
```{r}
tmap_save(map_mean_ai, 'mean_ai_lsoa.png')
```

- add inner outer borough character for each lsoa 

```{r}
CAL_and_PTAL_lsoa<-CAL_and_PTAL_lsoa%>%
  left_join(., st_drop_geometry(lsoa_new), by = join_by('code'=='code'))

innerouter <- london%>%
  select("name","inner_outer")

CAL_and_PTAL_lsoa<-CAL_and_PTAL_lsoa%>%
  left_join(., st_drop_geometry(innerouter), by = join_by('LAD11NM'=='name'))

CAL_and_PTAL_lsoa<-CAL_and_PTAL_lsoa%>%
  unique()

```

- average AI by inner outer

```{r}
mean_ai_inner_outer <- CAL_and_PTAL_lsoa %>%
  group_by(inner_outer) %>%
  summarise(mean=mean(average_AI, na.rm=TRUE))

mean_ai_inner_outer
```
```{r}
summary(CAL_and_PTAL_lsoa)
```
- box plot by inner outer and by borough 

```{r}
library(viridis)

box =
CAL_and_PTAL_lsoa %>%
  ggplot(aes(x=average_AI, y=reorder(LAD11NM, average_AI, FUN=mean),fill=inner_outer)) +
    geom_boxplot() +
    scale_fill_viridis(discrete = TRUE, begin=0.4, end=0.9, alpha=0.9,option="D") +
    theme_minimal() +
    theme(legend.position = "bottom", 
      plot.title = element_text(size=11)
    ) +
    ggtitle("Average AI per LSOA by borough") +
    xlab("Average AI")+
    ylab("")+
    labs(fill = "")


box
```
```{r}
ggsave("box_borough_lsoa.png", width= 10, height = 7, dpi = 600)
```

# Objective 3: identify areas of potential using data from PTAL and CAL 

This section will compare CAL and PTAL for the two case study boroughs through the bivariate choropleth maps to assess area of potential. 

```{r}
CAL_and_PTAL_borough_grid <-  st_join(CAL_and_PTAL_all, london)

CAL_and_PTAL_brent <- CAL_and_PTAL_borough_grid%>%
  filter(name == "Brent")
CAL_and_PTAL_brom <- CAL_and_PTAL_borough_grid%>%
  filter(name == "Bromley")
```

## plot the CAL and PTAL of both boroughs

```{r}
COLOR = c('white','#134480',"#1b71be","#3cb9e8", "#8dcc5b", "#ffeb3d", "#ffc08f", "#ff2e17","#8b1b16")
COLORcal = c('white',"#1b71be","#3cb9e8", "#8dcc5b", "#ffeb3d", "#ffc08f", "#ff2e17","#8b1b16")

tmap_mode("plot")

map_cal_brent =
  tm_shape(CAL_and_PTAL_brent) +
  tm_polygons(col = "CAL",palette=COLORcal,
          showNA = FALSE,
          colorNA = 'white',
          border.alpha = 0,
          title="") +
  tm_shape(brent) +
  tm_polygons(col = NA, 
              alpha = 0)+ 
tm_layout(bg.color = NA, 
          panel.labels = "Cycle Access Level",
          panel.label.size = 0.8,
          panel.label.bg.color = NA,
          frame = FALSE,
          legend.text.size = 0.7,
          legend.position = c("right","top"))+
  tm_compass(position=c(0,0.1)) + 
  tm_scale_bar(breaks = c(0, 2, 4),position=c(0,0.03)) +
  tm_credits("Data source: OSM, GLA, TfL. R package: r5r.", fontface = "italic", position=c(0,0))

map_ptal_brent =
  tm_shape(CAL_and_PTAL_brent) +
  tm_polygons(col = "PTAL2015",palette=COLOR,
          showNA = FALSE,
          colorNA = 'white',
          border.alpha = 0,
          title="") +
  tm_shape(brent) +
  tm_polygons(col = NA, 
              alpha = 0)+ 
tm_layout(bg.color = NA, 
          panel.labels = "Public Transport Access Level",
          panel.label.size = 0.8,
          panel.label.bg.color = NA,
          frame = FALSE,
          legend.text.size = 0.7,
          legend.position = c("right","top"))+
  tm_credits("Data source: TfL.", fontface = "italic", position=c(0,0))


t <- tmap_arrange(map_cal_brent, map_ptal_brent, ncol =2,asp = 1)
t
```

```{r}
tmap_save(t, 'brent_cal_ptal.png')
```

```{r}
tmap_mode("plot")

map_cal_brom =
  tm_shape(CAL_and_PTAL_brom) +
  tm_polygons(col = "CAL",palette=COLOR,
          showNA = FALSE,
          colorNA = 'white',
          border.alpha = 0,
          title="") +
  tm_shape(brom) +
  tm_polygons(col = NA, 
              alpha = 0)+ 
tm_layout(bg.color = NA, 
          panel.labels = "Cycle Access Level",
          panel.label.size = 0.8,
          panel.label.bg.color = NA,
            frame = FALSE,
           legend.text.size = 0.7,
            legend.position = c("right","bottom"))+
  tm_compass(position=c(0,0.1)) + 
  tm_scale_bar(breaks = c(0, 2, 4),position=c(0,0.03)) +
  tm_credits("Data source: OSM, GLA, TfL. R package: r5r.", fontface = "italic", position=c(0,0))

map_ptal_brom =
  tm_shape(CAL_and_PTAL_brom) +
  tm_polygons(col = "PTAL2015",palette=COLOR,
          showNA = FALSE,
          colorNA = 'white',
          border.alpha = 0,
          title="") +
  tm_shape(brom) +
  tm_polygons(col = NA, 
              alpha = 0)+ 
tm_layout(bg.color = NA, 
          panel.labels = "Public Transport Access Level",
          panel.label.size = 0.8,
          panel.label.bg.color = NA,
            frame = FALSE,
            legend.text.size = 0.7,
          legend.position = c("right","bottom"))+
  tm_credits("Data source: TfL.", fontface = "italic", position=c(0,0))


t2 <- tmap_arrange(map_cal_brom, map_ptal_brom, ncol =2,asp = 1)
t2
```

```{r}
tmap_save(t2, 'bromley_cal_ptal.png')
```

## bivariate choropleth maps

```{r}
# install.packages("biscale")
# install.packages("cowplot")
# install.packages("ggsn")
library(biscale)
library(cowplot)
library(ggsn)
```

```{r}
# make the process into function for easy reuse. 

bi_cho_data <- function (input) {
  
  # further group the CAL and PTAL bands into high, mid and low
  input2<-input%>%
    mutate(cal_num = case_when(CAL == "1a" | CAL == "1b" | CAL == "2"~ 1, 
                             CAL == "3" | CAL == "4"~ 2,
                             CAL == "5" | CAL == "6a" | CAL == "6b"~ 3,
                             TRUE ~ NA))%>%
    mutate(ptal_num = case_when(PTAL2015 == "1a" | PTAL2015 == "1b" | PTAL2015 == "2"~ 1, 
                             PTAL2015 == "3" | PTAL2015 == "4"~ 2,
                             PTAL2015 == "5" | PTAL2015 == "6a" | PTAL2015 == "6b"~ 3,
                             TRUE ~ NA))%>%
    filter(!is.na(cal_num) & !is.na(ptal_num))
  
  # specify data columns for mapping 
  data <- bi_class(input2, x = cal_num, y = ptal_num, style = "quantile", dim = 3)


  
  return(data)
}
```


```{r}
data_brent = bi_cho_data(CAL_and_PTAL_brent)  
```
```{r}
#add context (location in london)

st_crs(london) == st_crs(4326)

london_map <- ggplot()+
  geom_sf(data = london, fill='white', color = gray(.5))+ 
  theme_bw()+
  theme(panel.grid.major = element_line(color = gray(0.25), 
                                            linetype = "dashed", 
                                            linewidth = 0))+
  theme(panel.grid.major = element_blank(),
        panel.grid.minor = element_blank())+
  geom_sf(data = brent,aes(fill = 'red'), show.legend = FALSE)+
  theme(axis.title.x=element_blank(),
        axis.text.x=element_blank(),
        axis.ticks.x=element_blank(),
        axis.title.y=element_blank(),
        axis.text.y=element_blank(),
        axis.ticks.y=element_blank(),
        axis.line = element_blank(),
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
    panel.border = element_blank(),
    panel.background = element_rect(fill='transparent'),
    plot.background = element_rect(fill='transparent', color=NA)
    ) 

london_map
```

```{r}
#add context (key locations) + bi cho map

towncentres <- st_read(here::here("data/CentresPROJECTED.gpkg"))%>%
  select("objectid","borough","classification")%>%
  st_transform(4326)%>%
  filter(borough == "Brent" | borough == "Bromley")

tc_brent <- towncentres%>%
  filter(borough == "Brent")

tc_cen_brent <- towncentres_cen%>%
  st_join(., tc_brent)%>%
  filter(borough == "Brent")%>%
  select("id", "borough","classification","geometry")

# give classifcation of town centre a shape
tc_shape <- c('International'= 8, 'Metropolitan' = 21, 'Major' = 17, 'District' = 18,'Neighbourhood' = 20)
tc_col <- c('International'= "yellow", 'Metropolitan' = "yellow", 'Major' = "deeppink", 'District' = "darkgoldenrod1",'Neighbourhood' = "deepskyblue3")


# name of key town centres
tc_name <- tibble(name = c("Wembley", "Harlesden", "Old Oak Common", "Kilburn","Willesden", "Wembley Park","Preston Road"), Y = c(51.565, 51.525,51.526, 51.526, 51.560, 51.570, 51.575), X = c(-0.20, -0.22, -0.252, -0.19218331651448564, -0.198, -0.208, -0.217))

    
    
data_brent <- data_brent %>%st_transform(., 4326)

brent <- brent %>%st_transform(., 4326)

tc_cen_brent<- tc_cen_brent %>%st_transform(., 4326)

# create map
map <- ggplot() +
  coord_cartesian(xlim =c(-0.5, -0.10), ylim = c(51, 52)) +
  geom_sf(data = data_brent, mapping = aes(fill = bi_class), color = NA, size = 0.1, show.legend = FALSE) +
    bi_scale_fill(pal = "GrPink", dim = 3) +
    labs(title = "Bivariate choropleth map (CAL and PTAL)") +
    bi_theme()+
    theme(plot.title = element_text(size=8)) +
   geom_sf(data = brent, fill=alpha("white",0.2), color = 'black', lwd = .3, show.legend = FALSE)+ 
  geom_sf(data = tc_cen_brent, aes(shape = classification, col=classification), size =2, show.legend = FALSE)+
    scale_shape_manual(values=tc_shape)+
    scale_color_manual(values=tc_col)+
  draw_line(y = c(51.554721629396575, 51.565),
    x = c(-0.2860747704442487,-0.20),
    color = "midnightblue", size = .5, linetype=2)+
  draw_line(y = c(51.53603, 51.525),
    x = c(-0.2467083,-0.22),
    color = "midnightblue", size = .5, linetype=2)+
  draw_line(y = c(51.54741, 51.560),
    x = c(-0.2269566,-0.198),
    color = "midnightblue", size = .5, linetype=2)+
  draw_line(y = c(51.56154, 51.570),
    x = c(-0.2801694,-0.208),
    color = "midnightblue", size = .5, linetype=2)+
  draw_line(y = c(51.57188344664582,51.575),
    x = c(-0.29492742493101903,-0.217),
    color = "midnightblue", size = .5, linetype=2)+
  geom_label(data = tc_name, aes(X, Y, label = name), size = 2.5, show.legend = FALSE)+
   theme(axis.title.x=element_blank(),
        axis.text.x=element_blank(),
        axis.ticks.x=element_blank(),
        axis.title.y=element_blank(),
        axis.text.y=element_blank(),
        axis.ticks.y=element_blank(),
        axis.line = element_blank(),
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
    panel.border = element_blank(),
    panel.background = element_rect(fill='transparent'),
    plot.background = element_rect(fill='transparent', color=NA)
    ) 

map 

# legend
legend <- bi_legend(pal = "GrPink",
                    dim = 3,
                    xlab = "Higher CAL",
                    ylab = "Higher PTAL",
                    size = 4) +
   theme(panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
    panel.border = element_blank(),
    panel.background = element_rect(fill='transparent'),
    plot.background = element_rect(fill='transparent', color=NA)
    ) 
```


```{r}
# layout
finalPlot <- ggdraw() +
  draw_plot(map, 0.1, 0, 1, 1) +
  draw_plot(legend, .8, .65, .2, .2) +
  draw_plot(london_map, .8, .8, .2, .2) +
  theme(legend.box.margin = margin(0, 0, 0, 12))+
  theme(legend.key=element_rect(fill="white"))

finalPlot
```



```{r}
#add context (MCDA in previous file)

```

```{r}
ggsave("bi_cho_brent.png", width = 16, height = 10, units = "cm")
```

# same for Bromley

```{r}
data_brom = bi_cho_data(CAL_and_PTAL_brom)  
```

```{r}
#add context (location in london)

london_map_brom <- ggplot()+
  geom_sf(data = london, fill='white', color = gray(.5))+ 
  theme_bw()+
  theme(panel.grid.major = element_blank(),
        panel.grid.minor = element_blank())+
  geom_sf(data = brom,aes(fill = 'red'), show.legend = FALSE)+
  theme(axis.title.x=element_blank(),
        axis.text.x=element_blank(),
        axis.ticks.x=element_blank(),
        axis.title.y=element_blank(),
        axis.text.y=element_blank(),
        axis.ticks.y=element_blank(),
        axis.line = element_blank(),
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
    panel.border = element_blank(),
    panel.background = element_rect(fill='transparent'),
    plot.background = element_rect(fill='transparent', color=NA)
    ) 

london_map_brom
```

```{r}
#add context (key locations) + bi cho map

tc_brom <- towncentres%>%
  filter(borough == "Bromley")

tc_cen_brom <- towncentres_cen%>%
  st_join(., tc_brom)%>%
  filter(borough == "Bromley")%>%
  select("id", "borough","classification","geometry")

# name of key town centres
tc_name <- tibble(name = c("Bromley", "Orpington", "Beckenham", "Penge","Petts Wood", "West Wickham"), Y = c(51.45, 51.45,51.44, 51.45, 51.44, 51.37628), X = c(0.01577909, 0.1001473, -0.0244727, -0.05272902, 0.07553703, -0.06))

    
    
data_brom <- data_brom %>%st_transform(., 4326)

brom <- brom %>%st_transform(., 4326)

tc_cen_brom<- tc_cen_brom %>%st_transform(., 4326)

# create map
map_brom <- ggplot() +
  coord_cartesian() +
  geom_sf(data = data_brom, mapping = aes(fill = bi_class), color = NA, size = 0.1, show.legend = FALSE) +
    bi_scale_fill(pal = "GrPink", dim = 3) +
    labs(title = "Bivariate choropleth map (CAL and PTAL)") +
    bi_theme()+
    theme(plot.title = element_text(size=8)) +
   geom_sf(data = brom, fill=alpha("white",0.2), color = 'black', lwd = .3, show.legend = FALSE)+ 
  geom_sf(data = tc_cen_brom, aes(shape = classification, col=classification), size =2, show.legend = FALSE)+
    scale_shape_manual(values=tc_shape)+
    scale_color_manual(values=tc_col) +
  draw_line(y = c(51.40445, 51.45),
    x = c(0.01577909,0.01577909),
    color = "midnightblue", size = .5, linetype=2)+
  draw_line(y = c(51.37616, 51.45),
    x = c(0.1001473,0.1001473),
    color = "midnightblue", size = .5, linetype=2)+
  draw_line(y = c(51.40876, 51.44),
    x = c(-0.0244727,-0.0244727),
    color = "midnightblue", size = .5, linetype=2)+
  draw_line(y = c(51.41424, 51.45),
    x = c(-0.05272902,-0.05272902),
    color = "midnightblue", size = .5, linetype=2)+
  draw_line(y = c(51.38897,51.44),
    x = c(0.07553703,0.07553703),
    color = "midnightblue", size = .5, linetype=2)+
   draw_line(y = c(51.37628,51.37628),
    x = c(-0.01670207,-0.06),
    color = "midnightblue", size = .5, linetype=2)+
  geom_label(data = tc_name, aes(X, Y, label = name), size = 2.5, show.legend = FALSE)+
   theme(axis.title.x=element_blank(),
        axis.text.x=element_blank(),
        axis.ticks.x=element_blank(),
        axis.title.y=element_blank(),
        axis.text.y=element_blank(),
        axis.ticks.y=element_blank(),
        axis.line = element_blank(),
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
    panel.border = element_blank(),
    panel.background = element_rect(fill='transparent'),
    plot.background = element_rect(fill='transparent', color=NA)
    ) 

map_brom


# legend
legend_brom <- bi_legend(pal = "GrPink",
                    dim = 3,
                    xlab = "Higher CAL",
                    ylab = "Higher PTAL",
                    size = 4) +
   theme(panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
    panel.border = element_blank(),
    panel.background = element_rect(fill='transparent'),
    plot.background = element_rect(fill='transparent', color=NA)
    ) 
```


```{r}
# layout
finalPlot_brom <- ggdraw() +
  draw_plot(map_brom, 0.1, 0, 1, 1) +
  draw_plot(legend_brom, .8, .65, .2, .2) +
  draw_plot(london_map_brom, .8, .8, .2, .2) +
  theme(legend.box.margin = margin(0, 0, 0, 12))+
  theme(legend.key=element_rect(fill="white"))

finalPlot_brom
```

```{r}
ggsave("bi_cho_brom.png",width = 16, height = 10, units = "cm")
```

```{r}
tc_class = ggplot()+
    geom_sf(data = tc_cen_brom, aes(shape = classification, col=classification), size =2)+
 scale_shape_manual(values=tc_shape)+
    scale_color_manual(values=tc_col)+
  theme_dark()

legend_tc_class <- get_legend(tc_class)

legend_tc_class

grid.newpage()                              

grid.draw(legend_tc_class)

ggsave("tc_legend.png",width = 5, height = 10, units = "cm")

```


---
